task_manager_bot/requirements.txt
discord.py==2.3.2
pytest==8.3.5